move sidebar links to client component
